Remove number of messages
This information is superfluous.  What does it tell the reader?
If it doesn't fit on a page, there are more messages.  If it does
fit on a page, then it doesn't matter.  Also, the filtering does
not change the number of messages.  All in all, from a distraction
point of view, the actual number of messages is non-info.