Make Discord nicks filterable
Searchable is going to be a completely different matter, as that
would imply needing to pre-process entries in the IRC::Log state.
Which *may* be the ultimately best way to do it, but I'm not 100%
sure whether it would warrant changing the truthfulness of the
object representation of a log entry at that level.