Allow env to pass options in main program shebang line
When running the main program on Linux, `env` (which is used in the
main program's shebang line) throws this error:

```
$ ./bin/raku-logs-server
/usr/bin/env: ‘raku --ll-exception’: No such file or directory
/usr/bin/env: use -[v]S to pass options in shebang lines
```

In order to fix this issue, one needs to pass the [`-S` option to
`env`](https://www.gnu.org/software/coreutils/manual/html_node/env-invocation.html#g_t_002dS_002f_002d_002dsplit_002dstring-usage-in-scripts)
so that `env` can split the string passed to it and thus pass the
arguments to the interpreting program (i.e. `raku`).